Moved animation logic into SimpleMoveAnimation. Ghosting is still an issue.
