+++ src/UI/Art/SpriteArtist/PlayerSetupArtist.java
@@ -12,7 +12,7 @@ import UI.PlayerSetupController;
 import UI.PlayerSetupInfo;
 import UI.SlidingValue;
 import UI.UIUtils;
-import UI.Art.Animation.GameAnimation.AnimState;
+import UI.Art.SpriteArtist.UnitSpriteSet.AnimState;
 
 public class PlayerSetupArtist
 {
